Code cleanup and formatting
