+++ Source/Hurl.Selector/App.xaml.cs
@@ -1,13 +1,10 @@
 ﻿using Hurl.Selector.Pages;
 using Hurl.Selector.Services;
+using Hurl.Selector.Services.Interfaces;
 using Hurl.Selector.ViewModels;
 using Microsoft.Extensions.DependencyInjection;
 using System;
-using System.Diagnostics;
-using System.IO;
-using System.IO.Pipes;
 using System.Text.Json;
-using System.Threading;
 using System.Windows;
 
 namespace Hurl.Selector;
@@ -16,6 +13,8 @@ public partial class App : Microsoft.UI.Xaml.Application
 {
     public static IServiceProvider? Services { get; private set; }
 
+    private MainWindow? _mainWindow;
+
     public App()
     {
         Services = ConfigureServices();
@@ -23,7 +22,6 @@ public partial class App : Microsoft.UI.Xaml.Application
         Current.UnhandledException += Dispatcher_UnhandledException;
         DispatcherShutdownMode = Microsoft.UI.Xaml.DispatcherShutdownMode.OnExplicitShutdown;
     }
-   
 
     private static ServiceProvider ConfigureServices()
     {
@@ -37,11 +35,16 @@ public partial class App : Microsoft.UI.Xaml.Application
         return services.BuildServiceProvider();
     }
 
-    private MainWindow? _mainWindow;
+    protected override void OnLaunched(Microsoft.UI.Xaml.LaunchActivatedEventArgs args)
+    {
+        //var cliArgs = CliArgs.GatherInfo(args.Arguments, false);
+        //OpenedUri.Value = cliArgs.Url;
 
-    //private readonly CancellationTokenSource _cancelTokenSource = new();
-    //private Thread? _pipeServerListenThread;
-    //private NamedPipeServerStream? _pipeserver;
+        _mainWindow = new();
+        _mainWindow.SetContent(new SelectorPage());
+        //_mainWindow.Init(cliArgs);
+        _mainWindow.Show();
+    }
 
     private void Dispatcher_UnhandledException(object sender, Microsoft.UI.Xaml.UnhandledExceptionEventArgs e)
     {
@@ -62,62 +65,6 @@ public partial class App : Microsoft.UI.Xaml.Application
         string errorMessage = string.Format("{0}\n{1}\n\n{2}", ErrorMsgBuffer, e.Exception?.InnerException?.Message, e.Exception?.Message);
         MessageBox.Show(errorMessage, ErrorWndTitle, MessageBoxButton.OK, MessageBoxImage.Error);
     }
-
-    protected override void OnLaunched(Microsoft.UI.Xaml.LaunchActivatedEventArgs args)
-    {
-        //_pipeServerListenThread = new Thread(PipeServer);
-        //_pipeServerListenThread.Start();
-
-        //var cliArgs = CliArgs.GatherInfo(args.Arguments, false);
-        //OpenedUri.Value = cliArgs.Url;
-
-        _mainWindow = new();
-        _mainWindow.SetContent(new SelectorPage());
-        //_mainWindow.Init(cliArgs);
-        _mainWindow.Show();
-    }
-
-    //protected override void Exit()
-    //{
-    //    _cancelTokenSource.Cancel();
-    //    _pipeServerListenThread?.Join();
-
-    //    _singleInstanceMutex?.Close();
-    //    _singleInstanceWaitHandle?.Close();
-    //    _pipeserver?.Dispose();
-
-    //    base.OnExit(e);
-    //}
-
-    //public void PipeServer()
-    //{
-    //    while (!_cancelTokenSource.Token.IsCancellationRequested)
-    //    {
-    //        _pipeserver = new("HurlSelectorNamedPipe", PipeDirection.InOut, 1, PipeTransmissionMode.Byte, PipeOptions.Asynchronous);
-
-    //        try
-    //        {
-    //            _pipeserver.WaitForConnectionAsync(_cancelTokenSource.Token).Wait();
-
-    //            using StreamReader sr = new(_pipeserver);
-    //            string args = sr.ReadToEnd();
-    //            string[] argsArray = JsonSerializer.Deserialize<string[]>(args) ?? [];
-    //            _mainWindow.Show();
-    //        }
-    //        catch (OperationCanceledException)
-    //        {
-    //            break;
-    //        }
-    //        catch (Exception e)
-    //        {
-    //            Debug.WriteLine($"Error in PipeServer: {e.Message}");
-    //        }
-    //        finally
-    //        {
-    //            _pipeserver.Dispose();
-    //        }
-    //    }
-    //}
 }
