+++ Source/Hurl.Selector/Models/Settings.cs
@@ -1,9 +1,6 @@
 using CommunityToolkit.Mvvm.ComponentModel;
-using Hurl.Library;
 using System;
-using System.Collections.Generic;
 using System.Collections.ObjectModel;
-using System.IO;
 using System.Text.Json.Serialization;
 
 namespace Hurl.Selector.Models;
