+++ Source/Hurl.Selector/Services/JsonFileService.cs
@@ -1,11 +1,8 @@
 ﻿using Hurl.Library;
 using Hurl.Selector.Models;
-using System;
-using System.Collections.Generic;
+using Hurl.Selector.Services.Interfaces;
 using System.Collections.ObjectModel;
 using System.IO;
-using System.Linq;
-using System.Text;
 using System.Text.Json;
 using System.Threading.Tasks;
